EDD removed... * group management improved (front + back) * whole adding group and assigning it to user feature works fine

+++ src/main/java/org/engineer/work/service/impl/GroupServiceImpl.java
@@ -1,16 +1,11 @@
 package org.engineer.work.service.impl;
 
 import org.engineer.work.dto.GroupDTO;
-import org.engineer.work.exception.group.GroupExistsException;
-import org.engineer.work.exception.group.GroupNotFoundException;
-import org.engineer.work.exception.user.UserNotFoundException;
 import org.engineer.work.model.GroupEntity;
 import org.engineer.work.model.UserEntity;
 import org.engineer.work.repository.GroupRepository;
 import org.engineer.work.service.GroupService;
 import org.engineer.work.service.UserService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,8 +20,6 @@ import java.util.List;
 @Service
 public class GroupServiceImpl implements GroupService {
 
-	private static final Logger LOG = LoggerFactory.getLogger(GroupServiceImpl.class);
-
 	@Autowired
 	private GroupRepository groupRepository;
 
@@ -35,41 +28,65 @@ public class GroupServiceImpl implements GroupService {
 
 	@Override
 	public GroupEntity getGroupByName(final String name) {
-		return groupRepository.findOne(name);
+		GroupEntity groupEntity = null;
+		if (name != null) {
+			groupEntity = groupRepository.findOne(name);
+		}
+		return groupEntity;
 	}
 
 	@Override
 	@Transactional
-	public void createGroup(final GroupDTO groupDTO) throws GroupExistsException {
-		if (!groupRepository.exists(groupDTO.getName())) {
-			groupRepository.save(new GroupEntity(groupDTO));
-		} else {
-			throw new GroupExistsException(groupDTO.getName());
+	public void updateGroupMembers(final String user, final String group) {
+		if (user != null && group != null) {
+			final UserEntity userEntity = userService.getUserByUsername(user);
+			final GroupEntity groupEntity = this.getGroupByName(group);
+
+			if (userEntity != null && groupEntity != null) {
+				final List<GroupEntity> userGroups = userEntity.getGroups();
+
+				if (!userGroups.contains(groupEntity)) {
+					userGroups.add(groupEntity);
+
+					userEntity.setGroups(userGroups);
+					userService.updateUser(userEntity);
+				}
+			}
 		}
 	}
 
 	@Override
+	@Transactional
+	public boolean createGroup(final GroupDTO groupDTO) {
+		boolean result = false;
+		if (groupDTO != null && !groupRepository.exists(groupDTO.getName())) {
+			groupRepository.save(new GroupEntity(groupDTO));
+			this.updateGroupMembers(groupDTO.getGroupOwner(), groupDTO.getName());
+			result = true;
+		}
+		return result;
+	}
+
+	@Override
 	public List<GroupEntity> getAllGroups() {
-		List<GroupEntity> list = new ArrayList<>();
+		final List<GroupEntity> list = new ArrayList<>();
 		groupRepository.findAll().iterator().forEachRemaining(list::add);
 
 		return list;
 	}
 
 	@Override
-	public void deleteGroup(final String name) throws GroupNotFoundException {
-		if (groupRepository.exists(name)) {
+	@Transactional
+	public boolean deleteGroup(final String name) {
+		boolean result = false;
+		if (name != null && groupRepository.exists(name)) {
 			for (final UserEntity userEntity : this.getGroupByName(name).getUsers()) {
 				userEntity.getGroups().removeIf(group -> group.getName().equals(name));
-				try {
-					userService.updateUser(userEntity);
-				} catch (UserNotFoundException e) {
-					LOG.warn("User could not be found", e);
-				}
+				userService.updateUser(userEntity);
 			}
 			groupRepository.delete(name);
-		} else {
-			throw new GroupNotFoundException(name);
+			result = true;
 		}
+		return result;
 	}
 }
