+++ src/main/java/org/engineer/work/service/impl/UserServiceImpl.java
@@ -1,8 +1,6 @@
 package org.engineer.work.service.impl;
 
 import org.engineer.work.dto.UserDTO;
-import org.engineer.work.exception.user.UserExistsException;
-import org.engineer.work.exception.user.UserNotFoundException;
 import org.engineer.work.model.UserEntity;
 import org.engineer.work.repository.UserRepository;
 import org.engineer.work.service.UserService;
@@ -25,29 +23,33 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserEntity getUserByUsername(final String username) {
-        return userRepository.findOne(username);
+        UserEntity userEntity = null;
+        if (username != null) {
+            userEntity = userRepository.findOne(username);
+        }
+        return userEntity;
     }
 
     @Override
     @Transactional
-    public void createUser(final UserDTO userDTO) throws UserExistsException {
-        final String username = userDTO.getUsername();
-
-        if (! userRepository.exists(username)) {
+    public boolean createUser(final UserDTO userDTO) {
+        boolean result = false;
+        if (userDTO != null && !userRepository.exists(userDTO.getUsername())) {
             userRepository.save(new UserEntity(userDTO));
-        } else {
-            throw new UserExistsException(username);
+            result = true;
         }
+        return result;
     }
 
     @Override
     @Transactional
-    public void updateUser(final UserEntity userEntity) throws UserNotFoundException {
-        if (userRepository.exists(userEntity.getUsername())) {
+    public boolean updateUser(final UserEntity userEntity) {
+        boolean result = false;
+        if (userEntity != null && userRepository.exists(userEntity.getUsername())) {
             userRepository.save(userEntity);
-        } else {
-            throw new UserNotFoundException(userEntity.getUsername());
+            result = true;
         }
+        return result;
     }
 
     @Override
@@ -61,11 +63,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void deleteUser(final String username) throws UserNotFoundException {
-        if (userRepository.exists(username)) {
+    public boolean deleteUser(final String username) {
+        boolean result = false;
+        if (username != null && userRepository.exists(username)) {
             userRepository.delete(username);
-        } else {
-            throw new UserNotFoundException(username);
+            result = true;
         }
+        return result;
     }
 }
